CDL2Resolve python script for importing CDL folder to resolve timeline
This is a very early test of importing a folder of CDL files to Davinci Resolve. Relies on specific naming conventions, namely YYYYMMDD_REEL_CLIP_LOOKNAME.CDL and matching luts.

Written with a python 3 environment setup for resolve, and on windows so likely some changes to make it cross compatible.